+++ lib/action.js
@@ -1,9 +1,3 @@
-// TODO: awful dependencies here. How can we decouple them?
-const {
-	setLocalStorage,
-} = require( './init.js' );
-const { getRootUrlFromEnv } = require( './misc.js' );
-
 /**
  * Create an action. This function is designed to be used in a action file to be read.
  * Upon writing this, a action file is read as a node module. Thus, the fundamental format
@@ -111,16 +105,6 @@ const isDone = ( { done } ) => done;
 
 const isEarlyReturn = ( r ) => isAbort( r ) || isDone( r );
 
-const findFirstPath = ( actions ) => {
-	for ( const action of actions ) {
-		if ( ! action.isPreparation && action.initialPath ) {
-			return action.initialPath;
-		}
-	}
-
-	return '/';
-}
-
 const runSingleAction_ = async (
 	{ browser, context, page, config },
 	action,
@@ -216,17 +200,6 @@ const runActions_ = async ( runEnv, actions, accum ) => {
  * @return {object} a chain-accumulated result of running the given sequence of actions.
  */
 const runActions = async ( runEnv, actions ) => {
-	const { page, config } = runEnv;
-
-	// the very beginning navigation so the following can start from a valid origin
-	const beginningPath = findFirstPath( actions );
-	await page.goto( getRootUrlFromEnv( config.env ) + beginningPath );
-
-	// Localstorage can't be set without a domain, hence putting it here.
-	if ( config.localStorage ) {
-		await setLocalStorage( page, config.localStorage );
-	}
-
 	return await runActions_( runEnv, actions, {} );
 };
