+++ lib/init.js
@@ -1,9 +1,10 @@
+const { getRootUrlFromEnv } = require( './misc.js' );
+
 const initialize = async ( config ) => {
 	const {
 		browser: browserSlug,
 		locale,
 		rootUrl,
-		cookies = [],
 	} = config;
 
 	const module = require( 'playwright' );
@@ -23,11 +24,6 @@ const initialize = async ( config ) => {
 		locale,
 	} );
 
-	await context.addCookies( [
-		...cookies,
-		...convertCookiesProps( config ),
-	] );
-
 	// config the page instance
 	const page = await context.newPage();
 
@@ -35,7 +31,36 @@ const initialize = async ( config ) => {
 		browser,
 		context,
 		page,
+		config,
 	};
+};
+
+const postFirstNavigationSetup = async ( { context, page, config } ) => {
+	let needReload = false;
+
+	const cookies = [
+		...( config.cookies ? config.cookies : [] ),
+		...convertCookiesProps( config ),
+	];
+
+	// it looks like context-wise cookies are enough for wp.com for now.
+	// when we need page-wise cookies, `needReload` will need to set as true for that.
+	if ( cookies.length !== 0 ) {
+		await context.addCookies( [
+			...cookies,
+			...convertCookiesProps( config ),
+		] );
+	}
+
+	if ( config.localStorage ) {
+		await setLocalStorage( page, config.localStorage );
+
+		needReload = true;
+	}
+
+	if ( needReload ) {
+		await page.reload();
+	}
 };
 
 const convertCookiesProps = ( { currency } ) => {
@@ -62,7 +87,25 @@ const setLocalStorage = async ( page, entries ) => {
 	}
 };
 
+const findFirstPath = ( actions ) => {
+	for ( const action of actions ) {
+		if ( ! action.isPreparation && action.initialPath ) {
+			return action.initialPath;
+		}
+	}
+
+	return '/';
+}
+
+const goToStartUrl = async ( runEnv, actions ) => {
+	const { page, config } = runEnv;
+	const beginningPath = findFirstPath( actions );
+
+	return await page.goto( getRootUrlFromEnv( config.env ) + beginningPath );
+};
+
 module.exports = {
-	setLocalStorage,
 	initialize,
+	goToStartUrl,
+	postFirstNavigationSetup,
 };
